Update: Added order legs support for alpaca broker, insights and baseStrategy
We now track the legs of orders and pass them back to the insight it belong too. This is because we need access to if the order already has a TP or SL set as we don’t want to send multiple exit orders. since we have the order IDs of the leg we can modify it or cancel it so that we can manually control the exit