Fix: Backtesting - Take profit and stoploss on orders
When backtesting we didnt change the orderside to show that the position was closed. this made vbt not capture the position closes from our tp/sl. now it should be fine as we are seeing more trade closes in our backtests.

tempFix: yahoo finance minute ohlc data.
when using minute data with yf the high, low and close value for each bar is the same. so when using limit orders for entry/ tp/ sl levels they will never or rarely execute. so we added a 0.01 % padding on the high and low value for limit entry orders. This can't really be fixed on our end so we will look to offer eod  as a data provider for backtesting - in my todo list