Child insight (#6)
* Adding Child insights that only activate when the Parent insight is filled - mostly for DCA strategies

* Add star history chart to README.md

* chore: Update dependencies for quantstats

* Fix: Issue where 2 insights couldnt close in the same timeframe

Sometimes when you have 2 or more insighs close at the same time the first one insight will consume both filled / close trade events from the broker.

Updated: partial close orders are now checked before the insight close.

Impov: after a backtest you can see where the VBT portfolio is saved to.  as its logged out

* Fix: Backtesting - Take profit and stoploss on orders

When backtesting we didnt change the orderside to show that the position was closed. this made vbt not capture the position closes from our tp/sl. now it should be fine as we are seeing more trade closes in our backtests.

tempFix: yahoo finance minute ohlc data.
when using minute data with yf the high, low and close value for each bar is the same. so when using limit orders for entry/ tp/ sl levels they will never or rarely execute. so we added a 0.01 % padding on the high and low value for limit entry orders. This can't really be fixed on our end so we will look to offer eod  as a data provider for backtesting - in my todo list

* Added: stackDCA to main.py

* Fix: Add insight_id and parent fields to safe_insights dictionary

* Updated: dynamicQuantityToRiskExecutor not checking if quantity is already set

* added get latest Quotes function to Apha model base class

* feat: Add PercentageDCALevels executor for adding DCA levels to insights

* removed: strategy

* removed: old alpha