feat: Added featured bar events, now you can run a strategy with multiple timeframe resolutions
by default, the BaseStrategy.tradeOnFeatureEvents is set to False as we don’t plan to include taking trades directly from a feature during the backtest. So, it is best to use your main timeframe still to place trades and use the feature to check the state of higher or lower timeframes.

If you still want to trade on feature events you will need to set BaseStrategy.tradeOnFeatureEvents to True (in the constructor or assign it directly before you run the strategy)  and when placing any trades, filter out the feature suffix from the symbol string (eg `"btc-usd.15Min".split(".")[0]` -> btc-usd )