v0.3.0 (#7)
* Refactor paperBroker to work with live sim data from yf (15min delay)
Refactor BaseBroker to include get_current_time method

The BaseBroker class has been refactored to include a new method called get_current_time. This method returns the current datetime. This change allows for easier access to the current time within the broker.

* Update: Added order legs support for alpaca broker, insights and baseStrategy

We now track the legs of orders and pass them back to the insight it belong too. This is because we need access to if the order already has a TP or SL set as we don’t want to send multiple exit orders. since we have the order IDs of the leg we can modify it or cancel it so that we can manually control the exit

* added the live paper trading test file. WIP: not every bar has data when using YF

* Fixed ALPACA Broker: get history for stock data after hours results in none so we have to catch for that.

* Fixed: Insight didnt correctly capture the legs when using the alpaca broker. on submit we need to capture the legs - for some reason it doesnt get the legs from the incoming trade event.
Added: you can now manually create a stop order with alpaca broker.
Refactor insight executors classes to handle managed legs - just dont try close the insight when it has legs attached to it.

* removed test from main

* Refactor BasicStopLossExecutor to use low and high prices for stop loss checks

* come better annotaion for zed ide

* Added better backtesting with Vector BT  (#4)

* Added: optional annotations in base files

* Added: Vector BT for the backtest

* Now using Poerty for package management - this is for pip deploy later on

* Fixed: RSI Divergance Alpha V0.2 - performance - capped

* simplified EMA crossover strat

* Added a simple way to analyse past backtests when you deside to save them
```py

strategy.run()
    if (strategy.MODE == IStrategyMode.BACKTEST):
        print(strategy.BACKTESTING_RESULTS)
        # get the first asset in the strategy and plot the backtesting results
        symbol = list(strategy.assets.keys())[0]
        print(f"Symbol: {symbol}")

        save_path = Path(f"backtests/{strategy.STRATEGY_ID}")
        save_path.mkdir(parents=True, exist_ok=True)

        if (strategy.BACKTESTING_RESULTS.get(symbol)):
            strategy.BACKTESTING_RESULTS[symbol].save(f"backtests/{strategy.STRATEGY_ID}/{symbol}-{strategy.resolution}-backtest")
            strategy.BACKTESTING_RESULTS[symbol].plot().show()
        else:
            save_path.rmdir()
            print("No backtesting results found")
```

* moved the alphas and executors into the start function of the strategy class. (set up so we can run the backtest from UI)

* added: strategy viewer on UI

* Updated ReadMe

* Changes: for better readability
BaseBroker.close_order -> BaseBroker.cancel_order

* moved executor and alpha models to the Strategy.start()

* FIXED: Issues with the backtesting - some orders wouldnt close the simulation position - thus it wouldn't log the exit signal for VBT.

#5

* Close #5 Fixed account balance issue

* Child insight (#6)

* Adding Child insights that only activate when the Parent insight is filled - mostly for DCA strategies

* Add star history chart to README.md

* chore: Update dependencies for quantstats

* Fix: Issue where 2 insights couldnt close in the same timeframe

Sometimes when you have 2 or more insighs close at the same time the first one insight will consume both filled / close trade events from the broker.

Updated: partial close orders are now checked before the insight close.

Impov: after a backtest you can see where the VBT portfolio is saved to.  as its logged out

* Fix: Backtesting - Take profit and stoploss on orders

When backtesting we didnt change the orderside to show that the position was closed. this made vbt not capture the position closes from our tp/sl. now it should be fine as we are seeing more trade closes in our backtests.

tempFix: yahoo finance minute ohlc data.
when using minute data with yf the high, low and close value for each bar is the same. so when using limit orders for entry/ tp/ sl levels they will never or rarely execute. so we added a 0.01 % padding on the high and low value for limit entry orders. This can't really be fixed on our end so we will look to offer eod  as a data provider for backtesting - in my todo list

* Added: stackDCA to main.py

* Fix: Add insight_id and parent fields to safe_insights dictionary

* Updated: dynamicQuantityToRiskExecutor not checking if quantity is already set

* added get latest Quotes function to Apha model base class

* feat: Add PercentageDCALevels executor for adding DCA levels to insights

* removed: strategy

* removed: old alpha

* chore: Update Docker guide in README.md for running the bot with UI option enabled

* chore: Update license to Apache 2.0 and include OlympusTrader package in pyproject.toml

* Added: ISupportedBrokerFeatures dataclass to each broker.

* feat: Added featured bar events, now you can run a strategy with multiple timeframe resolutions

by default, the BaseStrategy.tradeOnFeatureEvents is set to False as we don’t plan to include taking trades directly from a feature during the backtest. So, it is best to use your main timeframe still to place trades and use the feature to check the state of higher or lower timeframes.

If you still want to trade on feature events you will need to set BaseStrategy.tradeOnFeatureEvents to True (in the constructor or assign it directly before you run the strategy)  and when placing any trades, filter out the feature suffix from the symbol string (eg `"btc-usd.15Min".split(".")[0]` -> btc-usd )

* refactor: Update dynamicQuantityToRiskExecutor to handle maximum cost basis based on supported broker features

* Updated: v0.3.0