Mt5 support - for the dashboard  (#11)
* added WIP CCXT

* WIP: MT5 broker impl - needs to be tested

* FIXES:

	- MT5 login must be a int but reading from env caused it to be a
	  string
	- The on bar format from history

* You could also get None  too for position

* changed to 1 min tf for testing

* Fixes:
- Positions fetch
- Streaming Market data to not pass old bars to the callback -> strat

* WIP: Added backword compat+ for contract size  from quantity for insights. Will need to fix breaking changes when asset does not have a contract_size set.

Fixed: MT5  UTC time issue by fetching the symbol time and shifint local time by the off set.

WIP: I need to fix the order trade streaming - orders dont come through. This might be cause of the UTC time issue again and will need to be updated.
WIP: Need to fix the inital order submission from insights  - as it does not format to an order.

* FIXED: the time zone sync issue - we find the time offset and round to the nearest hour and create a timezone from that info. when we want to use it we chnage our current time by the timezone and /or cast the time that we see is already in UTC as we know what the time should be (our mins + timezone hour shift).

ADDED: Test Entry Alpha - you can toggle between limit entry or market entries - by default its set to market entries.

* Update: max test orders to create from the test alpha

The test alpha only created one orders at the same time. but now we can set the number orders that it can create. this can be used to test 2 or more orders at the same time

* Update dynamicQuantityToRisk.py

now changes the size_should_buy  to respect contract_size, if it is set and should be used.  We also cap to the max order size.

* Update insight.py

Improved the update quantity to contracts flow

* Update mt5brokertestv2.py

Enabled The new UI mode for live mode

* Update tradesTable.py

positions in MT5  are the Order ID so need to look for that to get the current price.